ファイルが処理されていなかったときの処理を追加

+++ move_serviceparts.py
@@ -8,6 +8,10 @@ load_dotenv()
 
 serviceparts_dir = Path(os.getenv('SERVICE_PARTS_DIR'))
 box_dir = Path(os.getenv('BOX_DIR'))
+file = serviceparts_dir / os.getenv('FILE_NAME')
 
-for file in serviceparts_dir.glob('*.txt'):
+# fileが存在していたらコピー (create_serviceparts_tsv_pyでSOKEN.txtに追加するためのフラグ)
+if file.exists():
+    shutil.copy(file, box_dir / file.name)
+else:
     shutil.move(file, box_dir / file.name)
